requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
exceptiongroup==1.3.0
fastapi==0.118.2
frozenlist==1.8.0
h11==0.16.0
idna==3.10
multidict==6.7.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==10.4.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
pyngrok==7.4.0
pypdfium2==4.30.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
ttkbootstrap==1.14.4
twilio==9.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.22.0
